Refactor the page into app structure
